restructuring folder structure. creating more abstract TextInput to replace name input
