+++ src/App.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import './App.css';
-import RefereeInput from './RefereeInput/RefereeInput';
+import RefereeInput from './Inputs/RefereeInput/RefereeInput';
 import { AppState } from './store/store';
 import { connect } from 'react-redux';
 import RefereeList from './RefereeList/RefereeList';
 import { ObjectState } from './store/objectReducer';
 import { Referee } from './types/referee';
+import GameInput from './Inputs/GameInput/GameInput';
 
 interface AppProps {
   referees: ObjectState<Referee>
@@ -19,6 +20,7 @@ const App: React.FC<AppProps> = () => {
     <div className="App">
       <RefereeInput></RefereeInput>
       <RefereeList></RefereeList>
+      <GameInput></GameInput>
     </div>
   );
 }
